+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [Feature] New share flow(with `sf#path=()&id=()key=`)
 - [FIX] Uploading share errors
 - [FIX] Display by white color meta data on NFC card
+- [FIX] Scrim status bar and design changes on bottom sheet
 
 # 1.3.0